Implement universal Send To Rapp Command
Add a Send To R.app command so that anytime we need to send R commands we
can use this command instead of writing out the osascript implementation.
Also modified current commands to take advantage of the new system.

Added a settings file with the ability to set the app to send commands to,
also if they should be split into individual lines, or submitted as one
chunk, like the TextMate R bundle.